Make paths compatible on unix without IOMap
When loading solutions and projects, paths are now expanded to the full
path name and have their directory separators replaced with the native
seperator if necessary. The former change makes the paths to source
files more compatible with vim, which always provides a file's full path
when requesting a buffer's name via the python api.