Remove incorrect version information from syntax
The omnisharp syntax file was copied from the standard vimfiles, but
it's had some heavy modifications despite "last change" being 2009.

Add OmniSharp as a Maintainer and strip out the version lines to help
others find their way back to this repo and make it clearer that the
file isn't stock.