Set fzf as selector_ui when available
The selector_ui is set by default when it exists, by checking for
"autoload/{selector}.vim" in the runtimepath. However fzf has been
updated and there is no longer an "autoload/fzf.vim", but rather an
"autoload/fzf/vim.vim". To get around this, use "plugin/{selector}.vim",
which is more reliable.