Prevent navigation causing changes to be lost
Previously `:e!` was used to GotoDefinition or navigate to FindSymbols
results etc., which meant that if 'nohidden' was set, any changes in the
current buffer were lost.

Additionally, if the current buffer had a relative pwd and
GotoDefinition navigated to a definition in the same file, the relative
pwd of the buffer would not match the absolute pwd returned from the
server, so the current buffer was always re-read from disk with `:e!`.
This also resulted in changes to the current buffer being lost - even
when 'hidden' was set.

This commit changes `:e!` to `:e`, and also corrects the buffer
comparison, ensuring that absolute paths are always used.

A try-catch is also introduced to give users with 'nohidden' a nicer
warning message than the stacktrace they would otherwise see when
attempting to navigate away from a buffer containing changes.

Closes #364