Improve test for UltiSnips
The existing test for the presence of ultisnips performed an `exists()`
on an ultisnips function. This will fail when the ultisnips autoload
script has not yet been run.

This version uses `globpath()` to find the `plugin/UltiSnips.vim` file
in the runtimepath, which is also how we test for the presence of the UI
selector plugins.