Addresses issue #404 Roslyn server not installing
Fixes issue with PowerShell ExecutionPolicy blocking the install of
omnisharp-roslyn server files. This commit adds a check in the
autoload/OmniSharp.vim script for the PS policy level and checks for
failure messages sent from the PS install script.

Changes to the omnisharp-manager.ps1 script include updating the default
install location to be in the $HOME\.omnisharp\omnisharp-roslyn folder
instead of C:\; changing the unzipping method based on PS version;
confirming download and unzipping of the server files; and also changes
the download mode to SilentlyContinue to speed up the download.

Readme added in installer directory for instuctions on how to alter the ExecutionPolicy
settings in Powershell.