A few updates in addressing issue #404
Changed the output of the powershell script to return 0 for success
and 1 for failure. Also piped some of the output of the script to null
to control the final output when the script is completed. Added exit
codes that reflect the console output. In powershell this can be checked
with $LASTEXITCODE.

Implemented some suggestions that were given for cleaner and more
consistent code. Updated to handle the new output style from the
powershell script.

Changed the README for the installer directory to be more generic to
the directory instead of solely focused on the powershell issue. It
still needs to be update to include info for the .sh script.