Refactor :OmniSharpHighlightTypes and add variable
Add a `g:OmniSharp_highlight_types` variable, which enables a BufEnter
autocmd to be created to highlight C# buffers. This is different to the
previous system, where `:OmniSharpHighlightTypes` was used a single time
to begin highlighting (the autocmd was created as part of the command).

The command is now much simpler and simply triggers a refresh of the
highlighting, and it can be used alone without the variable or autocmd.