Use original buffer for highlighting keywords
After the /highlight call has returned, the buffer may have changed,
so the original buffer should be parsed for keywords, not the current
buffer.

This means performing the buffer/span parsing on the python side, with
the python returning the parsed types, instead of python returning spans
to then be compared on the vim side.