Workaround Windows fzf utf-8 code action problems
When fzf is passed non-ascii characters in Windows, an encoding error
can occur, and the selected action returned to the plugin no longer
matches the original string. fzf returns only the action name, no index
etc., so this means an action cannot be recognised.

This change (Windows only) adds an index prefix to all action names,
allowing actions to then be recognised by the prefix, ignoring the rest
of the name, which may be been "corrupted".

The change _only_ affects Windows, and the prefix is _only_ added and
matched against when any of the action names contain non-ascii
characters.

closes #574