This changeset fixes syntax highlighting via text properties, if the text contains multi byte characters.
The prop_add api of vim expects the column to be the byte offset and
not the character count. So if a comment for example would contain
a multi byte character, for example a german umlaut, then the last
character of the comment would not be included in the text property.