Fixes the issue that only the first line of a multiline highlight has a text property set.
Since it seems impossible to get the total number of lines for a specific
buffer in vimscript, the current implementation clears the text
properties step by step with each new highlight.

When a Highlight spans multiple lines and a there is another highlight
after it, the text properties of the previous multiline highlight are
being cleared.