Initialize buffers instead of full server load
The OmniSharp-roslyn loading issues appear to stem from situations where
multiple buffer-requests are sent before the project is fully
initialized. This results in duplicate "miscellaneous" files, which are
not correctly promoted to "project" files.

Instead, this commit introduces a buffer "initialization" concept, where
an `/updatebuffer` request is always sent as the first request for a
buffer, and no further requests are sent for that buffer until a
response is received (indicating that the server has registered the
file, and there will be no duplicates created).