Tighten up running-job project matching
When opening a buffer, running jobs are checked to see if this buffer
should be assigned to one of them. This is done by checking all of the
projects for each job (when there is no .sln, this will just be the
directory) and comparing with the current file name.

However, if the current path is e.g.:

    /project.a.b/file.cs

and there is a job containing:

    /project.a/proj.csproj

then we were getting a false positive, as:

    stridx('/project.a.b/file.cs', '/project.a') == 0

This has been fixed by including the directory separator in the match:

    stridx('/project.a.b/file.cs', '/project.a/') == -1

Additionally, WSL paths are converted to the local format, ensuring that
Windows directory separators are not matched incorrectly.