Improve HTTP server detection
To detect whether a user is using an HTTP server in stdio mode, we
maintain a count of messages from the server which could not be parsed
as valid JSON, as the stdio server returns JSON messages and the HTTP
server does not.

However this can lead to false positives, if the stdio server
unexpectedly outputs a series of non-json messages.

This patch adds a `job.logsize` count of valid json messages received,
and the server is only stopped when the number of invalid json messages
received is over 10 AND the logsize is less than 10.

Fixes #631