Display signatures in OmniSharpFindMembers (stdio only)
We are using the `v2/codestructure` endpoint instead of the
`/currentfilemembersasflat` endpoint here. The DisplayName of each item
in the response (elements and their children) contains the signature but
not the output type nor the modifiers like static/public/override etc.
so we are simply copying the beginning of the line in order to include
these.