Download new server before removing the old one
Previously during server installation, we deleted the existing server
and _then_ downloaded/extracted the new one.

However, github sometimes is not available, or the expected release
assets are not available, in which case the download or extraction fail.

This change prevents the old server from being removed until the
download and extraction are complete, ensuring that a working server is
not removed unless a valid new server has been prepared.