Handle arrow-key completion navigation
When using arrow keys to navigate the popup completion menu, the
selection is not inserted into the buffer, which makes it hard to
request signature help or documentation to display in the completion
popup.

The commit uses the same technique we use to request signature help
already, where we insert a '(' at the end of the method name in the
buffer we pass to the server, so that OmniSharp-roslyn will give us
results. Here we pass the full method/property name to the server.

Fixes #653