Do not use running sln when a closer sln exists
When selecting the project or solution to use for a buffer, we currently
first check for running server instances which are a parent of the
current buffer.

However this means that nested solutions will not ever be selected,
after a solution in a parent directory is running.

This commit first checks for ancestor solutions, and only resorts to
selecting running solutions when no ancestor .sln is found, or as a tie
breakder when multiple solutions are found.