Fix vader tests
The current vader async-helper tests depend on passing callback
functions, in order to know when to stop waiting for the async request
to complete. However, this no longer works, as the new callback logic
returns the results of the request without performing side-effects such
as navigation or quickfix population.

These tests are now simply waited on with a call to `sleep 5`.