Refactor NavigateUp/Down to use standard locations
The NavigatUp/NavigateDown functions have, until now, had their own
navigation implementation: setting a mark, changing the cursor position.

However, as we have standard functions for handling locations, it makes
sense to use them, especially with the new optional user callback, as this
makes the callback usage more predictable. The callback will now be
passed a "location" object (which looks like a quickfix item) with the
same structure as the locations returned to GotoDefinition callbacks and
FindUsages callbacks etc.