Add ! and mappings for running tests with no build
Previously this was done using a variable, g:OmniSharp_runtests_nobuild.
However, using the ! makes the test commands more flexible, and allows
users to easily alternate between tests which do and do not include the
build step.

The default behaviour been changed to include the build. This reverts
the previous change, which was breaking, and makes the OmniSharp-vim
commands match `dotnet test`, which _does_ build by default.