Start the server with at least minimal logging
When running the server with loglevel "None", no project load status
messages are received. This change ensures that when loglevel is set to
"None", no loglevel flag is sent to the server, so it logs at its
default "Info" level and we receive project status info. We just don't
write any logs to disk.